✨ feat: Avatar 컴포넌트 common index export 추가 (#18)
