+++ src/components/common/avatar/avatar.constants.ts
@@ -1,0 +1,4 @@
+export const SIZE_STYLES: Record<'sm' | 'md', string> = {
+  sm: 'w-8 h-8 text-sm',
+  md: 'w-10 h-10 text-base',
+} as const
