+++ src/components/common/index.ts
@@ -4,3 +4,4 @@ export * from '@/components/common/modal'
 export * from '@/components/common/image-upload'
 export * from '@/components/common/card'
 export * from '@/components/common/badge'
+export * from '@/components/common/avatar'
